Explicit (uint32_t) for ccr->tlen and ccr->tzlen
Other than reading from le32dec(), ccr->tlen is only assigned a non-zero value
in ccache_entry.c's callback_addtrailer(), where it gets the value of buflen.
callback_addtrailer() in turn is only called from the multitape layer (as
callback_trailer()), which only occurs inside callback_file()'s
    if (buflen < MINCHUNK)
[which is 4096].

ccr->tzlen has a similar story; it is only assigned a non-zero value inside
ccache_entry's callback_addtrailer(), where is gets a value of
    zlen = buflen + (buflen >> 9) + 13;
which will be 4115 in the worst case.

+++ tar/ccache/ccache_write.c
@@ -86,8 +86,8 @@ callback_write_rec(void * cookie, uint8_t * s, size_t slen, void * rec)
 	le64enc(ccre.size, ccr->size);
 	le64enc(ccre.mtime, ccr->mtime);
 	le64enc(ccre.nch, ccr->nch);
-	le32enc(ccre.tlen, ccr->tlen);
-	le32enc(ccre.tzlen, ccr->tzlen);
+	le32enc(ccre.tlen, (uint32_t)ccr->tlen);
+	le32enc(ccre.tzlen, (uint32_t)ccr->tzlen);
 	le32enc(ccre.prefixlen, plen);
 	le32enc(ccre.suffixlen, slen - plen);
 	le32enc(ccre.age, ccr->age + 1);
